ai-backend/requirements.txt
flask
flask-cors
transformers
pillow
opencv-python
torch
torchvision